+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Unreleased
 * Modify `mentos.py` to run on Python 3.x instead of Python 2.7
 * Add `:timeout` keyword option to allow for configurabel timeouts
 * Add several Python 3.x versions to test matrix
+* Drop dependency on MultiJson
+* Update Pygments to 2.7.3
 
 Version 1.2.1 (2017/12/07)
 -----------------------------